requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
docker==7.1.0
psutil==5.9.0
requests>=2.32.0

# Google Calendar API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
